refactor(lint/noEMptyInterface): ignore empty interface that extends a type
