+++ crates/biome_js_analyze/tests/specs/suspicious/noEmptyInterface/invalid.ts
@@ -1,17 +1,1 @@
-interface Baz extends Foo {}
-
 interface Foo {}
-
-interface Foo extends Array<number> {}
-
-interface Foo extends Array<number | {}> {}
-
-interface Foo<T> extends Bar<T> {}
-
-declare module FooBar {
-  export interface Bar extends Baz {}
-}
-
-namespace Ns {
-  export interface Bar extends Baz {}
-}
