+++ crates/biome_js_analyze/tests/specs/suspicious/noEmptyInterface/valid.d.ts
