+++ crates/biome_js_analyze/tests/specs/suspicious/noEmptyInterface/valid.ts
@@ -1,6 +1,4 @@
-interface A extends B {
-  prop: number;
-}
+interface A extends B { prop: number; }
 
 // valid because extending multiple interfaces
 // can be used instead of a union type
@@ -13,4 +11,21 @@ declare module "external" {
   global {
     export interface App extends Services {}
   }
+}
+
+// Ignore all interfaces that extends a type
+interface Baz extends Foo {}
+
+interface Foo extends Array<number> {}
+
+interface Foo extends Array<number | {}> {}
+
+interface Foo<T> extends Bar<T> {}
+
+declare module FooBar {
+  export interface Bar extends Baz {}
+}
+
+namespace Ns {
+  export interface Bar extends Baz {}
 }
